Move bootstrap tests into integration test directory
Extract working-directory and VM-option helpers into a focused module so
their white-box coverage can live under tests without exposing a library API.
